backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
Django==4.2.13
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
Werkzeug==3.0.3
